observes\rever(back): #5993.17 big ints
- this reverts commit 37dc35cd0a12e0abe37b301dc76ab40df8c4f41c.
- warning and error on jobs due to big int changes

+++ observes/singer/target_redshift/target_redshift/data_types.py
@@ -28,7 +28,7 @@ class NullOrEmptyType(Exception):
 
 raw_type_map: Dict[str, RedshiftDataType] = {
     "boolean": RedshiftDataType.BOOLEAN,
-    "integer": RedshiftDataType.BIGINT,
+    "integer": RedshiftDataType.INTEGER,
     "number": RedshiftDataType.DOUBLE_PRECISION,
     "string": RedshiftDataType.VARCHAR,
 }
